K8S/requirements.txt
fastapi
uvicorn[standard]
torch
transformers
Pillow
requests